+++ R/compile-layout.R
@@ -95,7 +95,7 @@ NULL
   labs <- .guide_labels(g)
   lw <- .mm_tw(labs, m$text_fs)
   body <- if (g$kind == "color_continuous") {
-    m$pad + m$bar_w + m$lab_gap + lw
+    m$pad + .bar_w_mm(g, m) + m$lab_gap + lw
   } else {
     m$pad + .guide_key_d(g, m) + m$lab_gap + lw
   }